# backend/docs/swagger.yaml
definitions:
  gin_http.ResponseJSON:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
info:
  contact: {}
paths:
  /admin/menu/list:
    get:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 获取后台管理菜单
...
  /api/v1/user/{id}:
    get:
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      summary: 获取用户信息
...
  /user/{id}:
    delete:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin_http.ResponseJSON'
      summary: 注销用户
  /user/{id}/password:
    put:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: password
        in: formData
        name: password
        required: true
        type: string
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin_http.ResponseJSON'
      summary: 修改用户密码
  /user/{id}/refreshToken:
    post:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: uuid
        in: formData
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin_http.ResponseJSON'
      summary: 更新Token
  /user/{id}/state:
    put:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: state
        in: query
        name: state
        required: true
        type: integer
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin_http.ResponseJSON'
      summary: 更新用户
  /user/login:
    post:
      parameters:
      - description: username
        in: formData
        name: username
        required: true
        type: string
      - description: password
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin_http.ResponseJSON'
        "20005":
          description: ""
          schema:
            $ref: '#/definitions/gin_http.ResponseJSON'
      summary: 用户登录
  /user/register:
    post:
      parameters:
      - description: username
        in: formData
        name: username
        required: true
        type: string
      - description: password
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin_http.ResponseJSON'
      summary: 用户注册
  /users:
    get:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin_http.ResponseJSON'
      summary: 用户列表
swagger: "2.0"